+++ 2 course/2 semester/Algorithms/lab2/algorithms_lab2_dijkstra/shortestPathFinding.cpp
@@ -1,7 +1,7 @@
 #include "shortestPathFinding.h"
 
 // Finds shortest path in graph using Dijkstra's algorithm with d-heap priority queue.
-void ldgDijkstraDHeap(std::vector<ElementOfAdjacencyList*> &ADJ, std::vector<unsigned long> &dist,
+void ldgDijkstraDHeap(std::vector<ElementOfAdjacencyList*> &ADJ, std::vector<unsigned long long> &dist,
 	std::vector<unsigned long> &up, GraphParameters *graphParameters)
 {
 	up.resize(graphParameters->numberOfVertices + 1);
@@ -10,7 +10,7 @@ void ldgDijkstraDHeap(std::vector<ElementOfAdjacencyList*> &ADJ, std::vector<uns
 	// Array of vertices' names. Name of vertex equals its index.
 	std::vector<unsigned int> name(graphParameters->numberOfVertices + 1);
 	// Represents current estimation of distance from vertex 's' to vertex 'name[i]'.
-	std::vector<unsigned long> key(graphParameters->numberOfVertices + 1);
+	std::vector<unsigned long long> key(graphParameters->numberOfVertices + 1);
 	// Represents names' indices. Is supposed to be correct at that: index[name[i]] = i
 	std::vector<unsigned int> index(graphParameters->numberOfVertices + 1);
 
@@ -18,17 +18,17 @@ void ldgDijkstraDHeap(std::vector<ElementOfAdjacencyList*> &ADJ, std::vector<uns
 	for (int i = 1; i <= graphParameters->numberOfVertices; i++)
 	{
 		up[i] = 0;
-		dist[i] = ULONG_MAX;
+		dist[i] = ULLONG_MAX;
 		index[i] = i;
 		name[i] = i;
-		key[i] = ULONG_MAX;
+		key[i] = ULLONG_MAX;
 	}
 
 	key[graphParameters->startingVertex] = 0;
 	int nq = graphParameters->numberOfVertices;
 	PriorityQueue priorityQueue(index, name, key, nq, graphParameters->d);
 	unsigned int name1 = name[1];
-	unsigned long key1 = key[1];
+	unsigned long long key1 = key[1];
 
 	while (nq > 0)
 	{
@@ -40,7 +40,7 @@ void ldgDijkstraDHeap(std::vector<ElementOfAdjacencyList*> &ADJ, std::vector<uns
 		{
 			unsigned int j = p->name;
 			unsigned int jq = index[j];
-			if (dist[i] != ULONG_MAX)
+			if (dist[i] != ULLONG_MAX)
 			{
 				if (key[jq] > dist[i] + p->weight)
 				{
