+++ 2 course/2 semester/Algorithms/lab2/algorithms_lab2_dijkstra/shortestPathFinding.h
@@ -3,5 +3,5 @@
 #include "structuresAndConstants.h"
 #include "priotityQueue.h"
 
-void ldgDijkstraDHeap(std::vector<ElementOfAdjacencyList*> &ADJ, std::vector<unsigned long> &dist,
+void ldgDijkstraDHeap(std::vector<ElementOfAdjacencyList*> &ADJ, std::vector<unsigned long long> &dist,
 	std::vector<unsigned long> &up, GraphParameters *graphParameters);
